requirements_traffic.txt
kafka-python
mysql-connector-python
numpy